Genrified classes for better typesafety and extensibility
