--- Java/src/main/java/com/evolvetechnology/TollCalculation.java
@@ -1,10 +1,0 @@
-package com.evolvetechnology;
-
-import com.evolvetechnology.vehicles.Vehicle;
-
-import java.time.LocalDateTime;
-
-@FunctionalInterface
-public interface TollCalculation {
-  int calculate(Vehicle vehicle, LocalDateTime... dates);
-}
